--- requirements.txt
@@ -1,11 +1,0 @@
-#
-# requirements.txt 
-#
-# Useful for configuring a python environment to run epanet-nrtestsuite. 
-#
-# command: pip install -r requirements.txt
-#
-
--e ./tools/epanet-output
--e ./tools/epanet-reader
--e ./tools/nrtest-epanet
