+++ tools/epanet-output/src/epanet_output.i
@@ -1,19 +1,22 @@
-%module outputapi
+/*
+ *  epanet_output.i - SWIG interface description file for EPANET Output API
+ * 
+ *  Created:    9/20/2017
+ *
+ *  Author:     Michael E. Tryby
+ *              US EPA - ORD/NRMRL
+ *
+*/ 
+%module epanet_output
 %{
 #include "errormanager.h"
 #include "messages.h"
-#include "outputapi.h"
+#include "epanet_output.h"
 
 #define SWIG_FILE_WITH_INIT
 %}
 
 %include "typemaps.i"
-%include "cstring.i"
-%include "numpy.i"
-
-%init %{
-    import_array();
-%}
 
 /* DEFINE AND TYPEDEF MUST BE INCLUDED */
 #define MAXMSG 53
@@ -25,6 +28,7 @@ typedef enum {
     ENR_link = 2
 } ENR_ElementType;
 
+/*
 typedef enum {
     ENR_nodeCount  = 1,
     ENR_tankCount  = 2,
@@ -32,6 +36,7 @@ typedef enum {
     ENR_pumpCount  = 4,
     ENR_valveCount = 5
 } ENR_ElementCount;
+*/
 
 typedef enum {
     ENR_flowUnits   = 1,
@@ -65,18 +70,13 @@ typedef enum {
 
 #ifdef WINDOWS
   #ifdef __cplusplus
-  #define DLLEXPORT extern "C" __declspec(dllexport) __stdcall
+  #define DLLEXPORT __declspec(dllexport) __cdecl
   #else
   #define DLLEXPORT __declspec(dllexport) __stdcall
   #endif
 #else
-  #ifdef __cplusplus
-  #define DLLEXPORT extern "C"
-  #else
   #define DLLEXPORT
-  #endif
 #endif
-
 
 /* TYPEMAPS FOR OPAQUE POINTER */
 /* Used for functions that output a new opaque pointer */
@@ -99,6 +99,7 @@ and return a (possibly) different pointer */
 /* TYPEMAP FOR IGNORING INT ERROR CODE RETURN VALUE */
 %typemap(out) int {
     $result = Py_None;
+    Py_INCREF($result);
 }
 
 /* TYPEMAPS FOR INT ARGUMENT AS RETURN VALUE */
@@ -109,39 +110,69 @@ and return a (possibly) different pointer */
     %append_output(PyInt_FromLong(*$1));
 }
 
-/* TYPEMAP FOR MEMORY MANAGEMENT OF STRINGS */
-%cstring_output_allocate(char** string_out, free(*$1))
+/* TYPEMAP FOR MEMORY MANAGEMENT AND ENCODING OF STRINGS */
+%typemap(in, numinputs=0)char** string_out (char* temp), int* slen (int temp){
+   $1 = &temp;
+}
+%typemap(argout)(char** string_out, int* slen) {
+    if (*$1) {
+        PyObject* o;
+        o = PyUnicode_FromStringAndSize(*$1, *$2);
+        
+        $result = SWIG_Python_AppendOutput($result, o);
+        free(*$1);
+    }
+}
 
-/* TYPEMAPS FOR MEMORY MANAGEMNET OF ARRAYS */
+/* TYPEMAPS FOR MEMORY MANAGEMNET OF FLOAT ARRAYS */
 %typemap(in, numinputs=0)float** float_out (float* temp), int* int_dim (int temp){
    $1 = &temp;
 }
 %typemap(argout) (float** float_out, int* int_dim) {
     if (*$1) {
-	  PyObject *o = PyList_New(*$2);
+      PyObject *o = PyList_New(*$2);
       int i;
       float* temp = *$1;
       for(i=0; i<*$2; i++) {
-        PyList_SetItem(o, i, PyFloat_FromDouble(temp[i]));
+        PyList_SetItem(o, i, PyFloat_FromDouble((double)temp[i]));
       }
       $result = SWIG_Python_AppendOutput($result, o);
-	  free(*$1);
+      free(*$1);
+    }
+}
+
+/* TYPEMAPS FOR MEMORY MANAGEMENT OF INT ARRAYS */
+%typemap(in, numinputs=0)int** int_out (long* temp), int* int_dim (int temp){
+   $1 = &temp;
+}
+%typemap(argout) (int** int_out, int* int_dim) {
+    if (*$1) {
+      PyObject *o = PyList_New(*$2);
+      int i;
+      long* temp = *$1;
+      for(i=0; i<*$2; i++) {
+        PyList_SetItem(o, i, PyInt_FromLong(temp[i]));
+      }
+      $result = SWIG_Python_AppendOutput($result, o);
+      free(*$1);
     }
 }
 
 /* TYPEMAP FOR ENUMERATED TYPES */
-%typemap(in) EnumeratedType (int val, int ecode = 0, PyObject * obj = 0) {
+%typemap(in) EnumeratedType (int val, int ecode = 0) {
     if (PyObject_HasAttrString($input,"value")) {
-    	obj = PyObject_GetAttrString($input, "value");
-      	ecode = SWIG_AsVal_int(obj, &val); 
-    } 	
+        PyObject* o;
+        o = PyObject_GetAttrString($input, "value");
+        ecode = SWIG_AsVal_int(o, &val); 
+    }   
     else {
         SWIG_exception_fail(SWIG_ArgError(ecode), "in method '" "$symname" "', argument " "$argnum"" of type '" "$ltype""'"); 
-    }  	
+    }   
     
     $1 = ($1_type)(val);
 }
-%apply EnumeratedType {ENR_ElementType, ENR_ElementCount, ENR_Units, ENR_Time, ENR_NodeAttribute, ENR_LinkAttribute}
+%apply EnumeratedType {ENR_ElementType, ENR_Units, ENR_Time, ENR_NodeAttribute, ENR_LinkAttribute}
+
 
 /* RENAME FUNCTIONS PYTHON STYLE */
 %rename("%(undercase)s") "";
@@ -159,39 +190,32 @@ and return a (possibly) different pointer */
     }
 }
 /* INSERT EXCEPTION HANDLING FOR THESE FUNCTIONS */  
-int DLLEXPORT ENR_open(ENR_Handle p_handle_in, const char* path);
+int DLLEXPORT ENR_open(ENR_Handle p_handle, const char* path);
 
-int DLLEXPORT ENR_getVersion(ENR_Handle p_handle_in, int* int_out);
-
-int DLLEXPORT ENR_getNetSize(ENR_Handle p_handle_in, ENR_ElementCount t_enum, int* int_out);
-
-int DLLEXPORT ENR_getUnits(ENR_Handle p_handle_in, ENR_Units t_enum, int* int_out);
-
-int DLLEXPORT ENR_getTimes(ENR_Handle p_handle_in, ENR_Time t_enum, int* int_out);
-
-int DLLEXPORT ENR_getElementName(ENR_Handle p_handle_in, ENR_ElementType t_enum,
-		int elementIndex, char** string_out);
-
-int DLLEXPORT ENR_getEnergyUsage(ENR_Handle p_handle_in, int pumpIndex,
+int DLLEXPORT ENR_getVersion(ENR_Handle p_handle, int* int_out);
+int DLLEXPORT ENR_getNetSize(ENR_Handle p_handle, int** int_out, int* int_dim);
+int DLLEXPORT ENR_getUnits(ENR_Handle p_handle, ENR_Units t_enum, int* int_out);
+int DLLEXPORT ENR_getTimes(ENR_Handle p_handle, ENR_Time t_enum, int* int_out);
+int DLLEXPORT ENR_getElementName(ENR_Handle p_handle, ENR_ElementType t_enum,
+		int elementIndex, char** string_out, int* slen);
+int DLLEXPORT ENR_getEnergyUsage(ENR_Handle p_handle, int pumpIndex,
 		int* int_out, float** float_out, int* int_dim);
+int DLLEXPORT ENR_getNetReacts(ENR_Handle p_handle, float** float_out, int* int_dim);
 
-int DLLEXPORT ENR_getNetReacts(ENR_Handle p_handle_in, float** float_out, int* int_dim);
 
-int DLLEXPORT ENR_getNodeAttribute(ENR_Handle p_handle_in, int periodIndex,
-        ENR_NodeAttribute t_enum, float** ARGOUTVIEWM_ARRAY1, int* DIM1);
-
-int DLLEXPORT ENR_getLinkAttribute(ENR_Handle p_handle_in, int periodIndex,
-        ENR_LinkAttribute t_enum, float** ARGOUTVIEWM_ARRAY1, int* DIM1);
+int DLLEXPORT ENR_getNodeAttribute(ENR_Handle p_handle, int periodIndex,
+        ENR_NodeAttribute t_enum, float** float_out, int* int_dim);
+int DLLEXPORT ENR_getLinkAttribute(ENR_Handle p_handle, int periodIndex,
+        ENR_LinkAttribute t_enum, float** float_out, int* int_dim);
 %exception;        
 
 /* NO EXCEPTION HANDLING FOR THESE FUNCTIONS */        
 int DLLEXPORT ENR_init(ENR_Handle* p_handle_out);        
-
 int DLLEXPORT ENR_close(ENR_Handle* p_handle_out);
+void DLLEXPORT ENR_free(void** array);
 
-void DLLEXPORT ENR_clearError(ENR_Handle p_handle_in);
-
-int DLLEXPORT ENR_checkError(ENR_Handle p_handle_in, char** msg_buffer);
+void DLLEXPORT ENR_clearError(ENR_Handle p_handle);
+int DLLEXPORT ENR_checkError(ENR_Handle p_handle, char** msg_buffer);
 
 
 /* CODE ADDED DIRECTLY TO SWIGGED INTERFACE MODULE */
@@ -201,13 +225,6 @@ import enum
 class ElementType(enum.Enum):
     NODE = ENR_node
     LINK = ENR_link
-
-class ElementCount(enum.Enum):
-    NODE_COUNT = ENR_nodeCount
-    TANK_COUNT = ENR_tankCount
-    LINK_COUNT = ENR_linkCount
-    PUMP_COUNT = ENR_pumpCount
-    VALVE_COUNT = ENR_valveCount
 
 class Units(enum.Enum):
     FLOW_UNIT = ENR_flowUnits
