+++ tools/epanet-output/test/data/__init__.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+
+#   __init__.py
+#
+#   Created: 11/13/2017
+#   Author: Michael E. Tryby
+#           US EPA - ORD/NRMRL
+#
+
+import os
+
+DATA_PATH = os.path.abspath(os.path.dirname(__file__))
+
+OUTPUT_FILE_EXAMPLE1 = os.path.join(DATA_PATH, 'net1.out')
