+++ tools/requirements.txt
@@ -1,0 +1,16 @@
+#
+# requirements.txt 
+#
+# Date Created: 10/10/2017
+# Author: Michael E. Tryby
+#         US EPA ORD/NRMRL
+#
+# Useful for configuring a python environment to run epanet-nrtestsuite. 
+#
+# command: 
+#  $ pip install --src build/packages -r tools/requirements.txt
+#
+
+-e  git+https://github.com/OpenWaterAnalytics/nrtest.git@master#egg=nrtest 
+-e ./tools/epanet-output
+-e ./tools/nrtest-epanet
