+++ tools/run-nrtest.sh
@@ -1,14 +1,21 @@
 #! /bin/bash
 
 #
-# run-nrtest.sh - Runs numerical regression test 
+#  run-nrtest.sh - Runs numerical regression test 
 #
-# Arguments:
-#  1 - test suite path
-#  2 - version/build identifier 
+#  Date Created: 10/16/2017
+#
+#  Author:       Michael E. Tryby
+#                US EPA - ORD/NRMRL
+#
+#  Arguments:
+#   1 - test suite path
+#   2 - version/build identifier 
 # 
+
 run-nrtest()
 {	
+
 return_value=0
 
 test_suite_path=$1
@@ -20,8 +27,8 @@ test_output_path="benchmark/epanet-$2"
 
 nrtest_compare_cmd="nrtest compare"
 ref_output_path="benchmark/epanet-2012"
-rtol_value=1.0
-atol_value=1.0
+rtol_value=0.1
+atol_value=0.0
 
 
 # change current directory to test_suite
@@ -30,15 +37,19 @@ cd ${test_suite_path}
 # clean test benchmark results
 rm -rf ${test_output_path}
 
-# python nrtest execute apps<app.json> tests<test.json> -o output
 echo INFO: Creating test benchmark
-${nrtest_execute_cmd} ${test_app_path} ${tests} -o ${test_output_path}
-
-echo 
-
-# python nrtest compare test_benchmark\ ref_benchmark\ --rtol --atol
-echo INFO: Comparing test and ref benchmarks
-return_value=$( ${nrtest_compare_cmd} ${test_output_path} ${ref_output_path} --rtol ${rtol_value} --atol ${atol_value} )
+nrtest_command="${nrtest_execute_cmd} ${test_app_path} ${tests} -o ${test_output_path}"
+echo INFO: "$nrtest_command"
+if ! [ $( $nrtest_command ) ]; then
+	echo  
+ 	echo INFO: Comparing test and ref benchmarks
+ 	nrtest_command="${nrtest_compare_cmd} ${test_output_path} ${ref_output_path} --rtol ${rtol_value} --atol ${atol_value}"
+ 	echo INFO: "$nrtest_command"
+ 	return_value=$( $nrtest_command ) 
+else
+ 	echo ERROR: Test benchmark creation failed
+ 	exit 1
+fi
 
 return $return_value
